requirements.txt
# Core dependencies
streamlit>=1.24.0
python-dotenv>=1.0.0
openai>=1.1.0
chromadb>=0.4.13
langchain>=0.0.267
langchain-openai>=0.0.2

# Document processing
pypdf>=3.15.1
pdfplumber>=0.10.2
tiktoken>=0.5.1

# Entity extraction (optional)
transformers>=4.30.2
torch>=2.0.1